backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
fastapi-mail
jinja2
pdfkit
qrcode
Pillow
python-dateutil
python-multipart
email-validator
python-slugify
openpyxl
reportlab
